refactor EventSearchRequestDto

+++ src/main/java/org/das/event_manager/dto/EventSearchRequestDto.java
@@ -15,7 +15,7 @@ public record EventSearchRequestDto(
       BigDecimal costMax,
       Integer durationMin,
       Integer durationMax,
-      Integer locationId,
+      Long locationId,
       EventStatus status
 ) {
 }
